+++ src/test/java/org/folio/login/controller/ApiExceptionHandlerTest.java
@@ -34,6 +34,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.ResponseEntity;
+import org.springframework.kafka.core.KafkaAdmin;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -45,6 +46,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @UnitTest
+@MockBean(KafkaAdmin.class)
 @WebMvcTest(ApiExceptionHandlerTest.TestController.class)
 @Import({ApiExceptionHandler.class, ApiExceptionHandlerTest.TestController.class})
 class ApiExceptionHandlerTest {